changed datatype of examId, activityId & subactivityId to long in entire affected source code

+++ app/src/main/java/in/principal/dao/ActivityMarkDao.java
@@ -5,7 +5,7 @@ import android.database.sqlite.SQLiteDatabase;
 
 public class ActivityMarkDao {
 
-    public static int getStudActMark(int studentId, int activityId, SQLiteDatabase sqliteDatabase){
+    public static int getStudActMark(int studentId, long activityId, SQLiteDatabase sqliteDatabase){
         int i = 0;
         Cursor c = sqliteDatabase.rawQuery("select Mark from activitymark where StudentId="+studentId+" and ActivityId="+activityId,null);
         c.moveToFirst();
@@ -17,7 +17,7 @@ public class ActivityMarkDao {
         return i;
     }
 
-    public static int getStudActAvg(int studentId, int activityId, SQLiteDatabase sqliteDatabase){
+    public static int getStudActAvg(int studentId, long activityId, SQLiteDatabase sqliteDatabase){
         int i = 0;
         Cursor c = sqliteDatabase.rawQuery("select (Avg(A.Mark)/B.MaximumMark)*100 as avg from activitymark A, activity B where A.ActivityId=B.ActivityId and A.ActivityId="+activityId+
                 " and StudentId="+studentId, null);
@@ -30,7 +30,7 @@ public class ActivityMarkDao {
         return i;
     }
 
-    public static int isThereActMark(int actId, int subjectId, SQLiteDatabase sqliteDatabase){
+    public static int isThereActMark(long actId, int subjectId, SQLiteDatabase sqliteDatabase){
         int isThere = 0;
         Cursor c = sqliteDatabase.rawQuery("select * from activitymark where ActivityId="+actId+" and SubjectId="+subjectId+" LIMIT 1", null);
         if(c.getCount()>0){
@@ -40,7 +40,7 @@ public class ActivityMarkDao {
         return isThere;
     }
 
-    public static int isThereActGrade(int actId, int subjectId, SQLiteDatabase sqliteDatabase){
+    public static int isThereActGrade(long actId, int subjectId, SQLiteDatabase sqliteDatabase){
         int isThere = 0;
         Cursor c = sqliteDatabase.rawQuery("select * from activitygrade where ActivityId="+actId+" and SubjectId="+subjectId+" LIMIT 1", null);
         if(c.getCount()>0){
